--HG--
branch : feature-light_cycle

+++ code/ryzom/server/src/admin_modules/aes_module.cpp
@@ -15,6 +15,7 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 
+#include "stdpch.h"
 #include "nel/misc/singleton.h"
 #include <time.h>
 #include "nel/misc/path.h"
@@ -573,7 +574,7 @@ namespace ADMIN
 						time_t t = now;
 						fprintf(fp, "AESReportDate=%s", ::ctime(&t));
 
-						fprintf(fp, "NBService=%u\n", _ServiceStates.size());
+						fprintf(fp, "NBService=%u\n", (uint)_ServiceStates.size());
 						// output state of each service
 						TServiceStates::iterator first(_ServiceStates.begin()), last(_ServiceStates.end());
 						for (; first != last; ++first)
